+++ tests/test_tokenizer.py
@@ -18,6 +18,10 @@ from test_keyvalues import parse_test as prop_parse_test
 T = Token
 IS_CPYTHON = platform.python_implementation() == 'CPython'
 
+# See https://github.com/ValveSoftware/source-sdk-2013/blob/0d8dceea4310fde5706b3ce1c70609d72a38efdf/sp/src/tier1/utlbuffer.cpp#L57-L69
+ESCAPE_CHARS = "\n \t \v \b \r \f \a \\ \' \""
+ESCAPE_ENCODED = r"\n \t \v \b \r \f \a \\ \' " + r'\"'
+
 # The correct result of parsing prop_parse_test.
 # Either the token, or token + value (which must be correct).
 prop_parse_tokens = [
@@ -239,7 +243,7 @@ def test_pushback(py_c_token: Type[Tokenizer]) -> None:
     tok = Tokenizer(prop_parse_test, '', string_bracket=True)
     tokens = []
     for i, (tok_type, tok_value) in enumerate(tok):
-        if i % 3 == 0:
+        if i % 5 in (0, 3, 4):
             tok.push_back(tok_type, tok_value)
         else:
             tokens.append((tok_type, tok_value))
@@ -248,7 +252,7 @@ def test_pushback(py_c_token: Type[Tokenizer]) -> None:
     tok = Tokenizer(noprop_parse_test, '')
     tokens = []
     for i, (tok_type, tok_value) in enumerate(tok):
-        if i % 3 == 0:
+        if i % 5 in (0, 3, 4):
             tok.push_back(tok_type, tok_value)
         else:
             tokens.append((tok_type, tok_value))
@@ -273,10 +277,9 @@ def test_pushback_opvalues(py_c_token: Type[Tokenizer], token: Token, val: str) 
     tok.push_back(token, val)
     assert tok() == (token, val)
 
-    # Can only push back one at a time.
     tok.push_back(Token.STRING, 'push')
-    with pytest.raises(ValueError, match='Token already pushed back!'):
-        tok.push_back(Token.STRING, 'two_at_once')
+    tok.push_back(Token.DIRECTIVE, 'second')
+    assert tok() == (Token.DIRECTIVE, 'second')
     assert tok() == (Token.STRING, 'push')
 
     # Value is ignored for these token types.
@@ -531,6 +534,8 @@ def test_obj_config(py_c_token: Type[Tokenizer], parm: str, default: bool) -> No
     ("\thello_world", r"\thello_world"),
     ("\\thello_world", r"\\thello_world"),
     ("\\ttest\nvalue\t\\r\t\n", r"\\ttest\nvalue\t\\r\t\n"),
+    (ESCAPE_CHARS, ESCAPE_ENCODED),
+    ('Unchanged ?/', 'Unchanged ?/'),
     # BMP characters, and some multiplane chars.
     ('test: ╒══╕', r'test: ╒══╕'),
     ("♜♞🤐♝♛🥌 chess: ♚♝♞♜", "♜♞🤐♝♛🥌 chess: ♚♝♞♜"),
@@ -680,10 +685,14 @@ def test_allow_escapes(py_c_token: Type[Tokenizer]) -> None:
         (Token.STRING, r"tab\ted"),
         Token.BRACE_CLOSE,
     ])
-    check_tokens(py_c_token(r'{ "string\n" "tab\ted q\tuo\"te\"" }', allow_escapes=True), [
+    check_tokens(py_c_token(
+        f'{{ "string\\n" "all escapes: {ESCAPE_ENCODED}" "also accepted /?" }}',
+        allow_escapes=True,
+    ), [
         Token.BRACE_OPEN,
         (Token.STRING, "string\n"),
-        (Token.STRING, 'tab\ted q\tuo\"te"'),
+        (Token.STRING, f'all escapes: {ESCAPE_CHARS}'),
+        (Token.STRING, 'also accepted /?'),
         Token.BRACE_CLOSE,
     ])
 
@@ -953,6 +962,9 @@ def test_subclass_base(Tok: Type[BaseTokenizer]) -> None:
     assert tok() == (Token.DIRECTIVE, '#test')
     assert tok() == (Token.NEWLINE, '\n')
     assert tok.peek() == (Token.BRACE_OPEN, '{')
+    tok.push_back(Token.STRING, "extra")
+    assert tok.peek() == (Token.STRING, "extra")
+    assert tok() == (Token.STRING, "extra")
     assert tok.peek() == (Token.BRACE_OPEN, '{')
     assert next(iter(tok)) == (Token.BRACE_OPEN, '{')
     skip = tok.skipping_newlines()